Configure `prettier` to run on commit via `husky`
This change sets up `lint-staged` and `husky`. `lint-staged` runs
`prettier` on a set of files that fit a given filename pattern. `husky`
allows us to configure Git hooks so that on commit, we can run
`lint-staged` scripts.

This way, when we run `git commit` in our code, any updated code will
automatically be formatted.